+++ generate/crypto_sign/cross-rsdp-256-fast/avx2/api.h
@@ -28,6 +28,7 @@
 
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
 
 #define PQCLEAN_CROSSRSDP256FAST_AVX2_CRYPTO_ALGNAME "cross-rsdp-256-fast"
@@ -49,30 +50,30 @@ int PQCLEAN_CROSSRSDP256FAST_AVX2_crypto_sign_keypair(unsigned char *pk,
                                                      );
 
 int PQCLEAN_CROSSRSDP256FAST_AVX2_crypto_sign(unsigned char *sm,
-        uint64_t *smlen,
+        size_t *smlen,
         const unsigned char *m,
-        uint64_t mlen,
+        size_t mlen,
         const unsigned char *sk
                                              );
 
 int PQCLEAN_CROSSRSDP256FAST_AVX2_crypto_sign_open(unsigned char *m,
-        uint64_t *mlen,
+        size_t *mlen,
         const unsigned char *sm,
-        uint64_t smlen,
+        size_t smlen,
         const unsigned char *pk
                                                   );
 
 int PQCLEAN_CROSSRSDP256FAST_AVX2_crypto_sign_signature(unsigned char *sig,
-        uint64_t *siglen,
+        size_t *siglen,
         const unsigned char *m,
-        uint64_t mlen,
+        size_t mlen,
         const unsigned char *sk
                                                        );
 
 int PQCLEAN_CROSSRSDP256FAST_AVX2_crypto_sign_verify(const unsigned char *sig,
-        uint64_t siglen,
+        size_t siglen,
         const unsigned char *m,
-        uint64_t mlen,
+        size_t mlen,
         const unsigned char *pk
                                                     );
